Move siren schema tests to dedicated file
